+++ lib/pages/buy_and_sell.dart
@@ -1,14 +1,18 @@
 /*
- * AUTOR:               Malo Perisé, Pedro
- * CREACIÓN:            12/03/2019
- * ÚLTIMA MODIFICACIÓN: 
- * DESCRIPCIÓN:         
+ * FICHERO:     buy_and_sell.dart
+ * DESCRIPCIÓN: clases relativas al la página de selección de compra/venta
+ * CREACIÓN:    13/03/2019
  */
-
 import 'package:flutter/material.dart';
 import 'package:bookalo/widgets/navbars/buy_and_sell_navbar.dart';
 import 'package:bookalo/pages/buy.dart';
+import 'package:bookalo/widgets/animations/bookalo_progress.dart';
 
+/*
+ *  CLASE:        BuyAndSell
+ *  DESCRIPCIÓN:  widget para la selección de bien la pestaña de compra,
+ *                bien la de venta. Permite deslizado de una a otra
+ */
 class BuyAndSell extends StatelessWidget {
 
   BuyAndSell();
@@ -23,7 +27,7 @@ class BuyAndSell extends StatelessWidget {
         body: TabBarView(
           children: [
             Buy(),
-            Icon(Icons.access_alarms),
+            BookaloProgressIndicator()
           ],
         ),
       ),
